Initial Uninitialized Stack Variable commit
Initial commit of the Uninitialized Stack Variable exploit and associated explanations. Pretty simple exploit to be honest, main thing is understanding how the stack works, the use of !thread to get the base of the thread's stack, and the use of undocumented functionality to gain control over the stack. Still a few things I could document r.e the undocumented functionality (whew that sounds ironic there) and the parameters we use though. Save that for later :)