Use After Free Inital Exploit Commit
Initial commit of Use After Free exploit with a ton of explanations and information surrounding the whats and whys. Hopefully will get a video done as well on this soon to assist. Should cover most of the questions, might still be one or two missing and there is one that is not 100% confirmed atm.